DOMA-3039 ported export ticket feature to worker in batch mode
