+++ apps/condo/domains/common/constants/worker.js
@@ -1,0 +1,5 @@
+const WORKER_FINGERPRINT = 'worker'
+
+module.exports = {
+    WORKER_FINGERPRINT,
+}
