+++ apps/condo/domains/ticket/utils/serverSchema/index.js
@@ -4,7 +4,6 @@
  * Please, don't remove `AUTOGENERATE MARKER`s
  */
 const { GqlWithKnexLoadList } = require('@condo/domains/common/utils/serverSchema')
-const compact = require('lodash/compact')
 const { generateServerUtils, execGqlWithoutAccess } = require('@condo/domains/common/utils/codegeneration/generate.server.utils')
 const { Ticket: TicketGQL } = require('@condo/domains/ticket/gql')
 const { TicketStatus: TicketStatusGQL } = require('@condo/domains/ticket/gql')
