+++ apps/condo/domains/ticket/utils/testSchema/index.js
@@ -10,7 +10,7 @@ const { makeClientWithProperty } = require('@condo/domains/property/utils/testSc
 const { makeLoggedInAdminClient } = require('@core/keystone/test.utils')
 const { TICKET_STATUS_TYPES, ORGANIZATION_COMMENT_TYPE } = require('../../constants')
 const { generateGQLTestUtils, throwIfError } = require('@condo/domains/common/utils/codegeneration/generate.test.utils')
-const { Ticket: TicketGQL } = require('@condo/domains/ticket/gql')
+const { Ticket: TicketGQL, EXPORT_TICKETS_TO_EXCEL } = require('@condo/domains/ticket/gql')
 const {
     TicketStatus: TicketStatusGQL,
     TicketChange: TicketChangeGQL,
@@ -32,6 +32,7 @@ const { UserTicketCommentReadTime: UserTicketCommentReadTimeGQL } = require('@co
 const { TicketPropertyHint: TicketPropertyHintGQL } = require('@condo/domains/ticket/gql')
 const { TicketPropertyHintProperty: TicketPropertyHintPropertyGQL } = require('@condo/domains/ticket/gql')
 const { ExportTicketTask: ExportTicketTaskGQL } = require('@condo/domains/ticket/gql')
+const { DEFAULT_ORGANIZATION_TIMEZONE } = require('@condo/domains/organization/constants/common')
 /* AUTOGENERATE MARKER <IMPORT> */
 
 const TICKET_OPEN_STATUS_ID ='6ef3abc4-022f-481b-90fb-8430345ebfc2'
@@ -629,7 +630,7 @@ async function createTestTicketExportTask (client, user, extraAttrs = {}) {
     return [obj, attrs]
 }
 
-async function updateTestTicketExportTask (client, id, extraAttrs = {}) {
+async function updateTestExportTicketTask (client, id, extraAttrs = {}) {
     if (!client) throw new Error('no client')
     if (!id) throw new Error('no id')
     const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
@@ -701,6 +702,7 @@ module.exports = {
     getTicketAnalyticsReport, getTicketAnalyticsExport,
     TicketPropertyHint, createTestTicketPropertyHint, updateTestTicketPropertyHint,
     TicketPropertyHintProperty, createTestTicketPropertyHintProperty, updateTestTicketPropertyHintProperty,
-    ExportTicketTask, createTestExportTicketTask, updateTestExportTicketTask,
+    ExportTicketTask, createTestTicketExportTask, updateTestExportTicketTask,
+    exportTestTicketsToExcel,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
